requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.19
Werkzeug==0.9.4
bottlenose==0.5.0
itsdangerous==0.23
lxml==3.3.3
pymongo==2.6.3
python-amazon-simple-product-api==1.4.0
wsgiref==0.1.2
